Issue/170 review score decimal conversion (#173)
* delete verify company button

* update riskScore intake type

* deleted component import

+++ adversarial_apps/src/app/api/call-python-api/python-calls/database.py
@@ -381,7 +381,7 @@ def verify_company(cik: str) -> dict:
             connection.close()
 
 
-def update_company_score(cik: str, risk_score: int) -> dict:
+def update_company_score(cik: str, risk_score: float) -> dict:
     """
     Update the risk score for a company based on the CIK number.
